# dist/api.yml
openapi: 3.0.0
info:
  title: api
  description: API for  project
  version: 1.0.0
...
paths:
...
  /users/lastSearch:
    get:
      tags:
        - Users
      summary: Return the query and result of the last search
      responses:
        "200":
          description: data about the last search
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/lastSearch"
        "400":
          description: wrong input parameters
...
  /users/favoritePlayers:
    get:
      tags:
        - Users
      summary: Return the favorite players that were saved by user
      parameters:
      responses:
        "200":
          description: data about the user's favorites players returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerPreview"
        "400":
          description: There is no team with this id inputs

    post:
      summary: save player in the favorites list of a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                playerId:
                  type: int
                  description: player's id which the user saved
                  example: 82808
      responses:
        "201":
          description: The player successfully saved as favorite
        "400":
          description: There is no player with this id
  /users/favoriteGames:
    get:
      tags:
        - Users
      summary: Return the favorite games that were saved by user
      parameters:
      responses:
        "200":
          description: data about the user's favorites games returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/gamesFutureDetails"
        "400":
          description: There are no games in favorites
    post:
      summary: save game in the favorites list of a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                gameId:
                  type: int
                  description: game's id which the user saved
                  example: 939
      responses:
        "201":
          description: The game successfully saved as favorite
        "400":
          description: incorrect inputs
  /users/favoriteTeams:
    get:
      tags:
        - Users
      summary: Return the favorite teams that were saved by user
      parameters:
      responses:
        "200":
          description: data about the user's favorites teams returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/teamInfo"
        "401":
          description: user is not authorized

    post:
      summary: save team in the favorites list of a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                teamId:
                  type: int
                  description: teams's id which the user saved
                  example: 939
      responses:
        "201":
          description: The team successfully saved as favorite
        "400":
          description: incorrect inputs
  /users/searchByName/{Name}:
    get:
      summary: get details of players, teams and coaches with the name
      description: useful for search page
      tags:
        - Users
      parameters:
        - in: path
          name: Name
          description: Name of the player/team/coach
          required: true
          schema:
            type: string
      responses:
        "200":
          description: data about the player/team/coach
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/search"
        "404":
          description: Name is not exists
  /users/searchByNameFilterWithPositionId/{Name}/{positionId}:
    get:
      summary: get details of players with the name and position Id
      description: useful for search page
      tags:
        - Users
      parameters:
        - in: path
          name: Name
          description: Name of the player
          required: true
          schema:
            type: string
        - in: path
          name: positionId
          description: position of the player
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: data about the player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/searchFilterWithPosition"
        "404":
          description: player with this name and position is not exists
  /users/searchByNameFilterWithTeamName/{Name}/{teamName}:
    get:
      summary: get details of players and coaches with the name and team name
      description: useful for search page
      tags:
        - Users
      parameters:
        - in: path
          name: Name
          description: Name of the player/coach
          required: true
          schema:
            type: string
        - in: path
          name: teamName
          description: team name of the player
          required: true
          schema:
            type: string
      responses:
        "200":
          description: data about the player/coach
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/searchFilterWithTeamName"
        "404":
          description: player or coach with this name and team name not exists
...
components:
  schemas:
...
    lastSearch:
      type: object
      properties:
        query:
          type: string
          description: the query of the search
          example: /searchByName/jess
        result:
          type: object
          $ref: "#/components/schemas/searchFilterWithTeamName"
    PlayerPreview:
      type: object
      properties:
        player_id:
          type: integer
          description: player's id
          example: 456421
        full_name:
          type: string
          description: player's full name
          example: Jess Thorup
        image:
          type: string
          description: player's image
          format: URL
          example: "https://cdn.sportmonks.com/images/soccer/placeholder.png"
        position:
          type: integer
          description: player's position in team
          example: 9
        team_name:
          type: string
          description: player's team name
          example: København
...
    CoachPreview:
      type: object
      properties:
        coach_id:
          type: integer
          description: coach's id
          example: 456421
        full_name:
          type: string
          description: coach's full name
          example: Jess Thorup
        team_name:
          type: string
          description: coach's team name
          example: København
        image:
          type: string
          description: coach's image
          format: URL
          example: null
...
    teamInfo:
      type: object
      properties:
        team_id:
          type: integer
          example: Midtjylland
          description: id of the team
        team_name:
          type: string
          example: Midtjylland
          description: the name of the team
        team_logo:
          type: string
          description: team's logo
          format: URL
          example: https://cdn.sportmonks.com/images//soccer/teams/11/939.png
    search:
      type: object
      properties:
        players:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/PlayerPreview"
        teams:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/teamInfo"
        coaches:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/CoachPreview"
    searchFilterWithPosition:
      type: array
      items:
        type: object
        $ref: "#/components/schemas/PlayerPreview"   
    searchFilterWithTeamName:
      type: object
      properties:
        players:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/PlayerPreview"
        coaches:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/CoachPreview"
...
    gamesFutureDetails:
      type: array
      items:
        type: object
        description: Contains future game
        properties:
          gameID:
            type: integer
            example: 12
            description: id of the game
          gamedate:
            type: date
            example: "2021-11-11"
            description: date
          gametime:
            type: string
            example: 20:00:00
            description: time
          hometeamID:
            type: integer
            example: 939
            description: home team id
          awayteamID:
            type: integer
            example: 1020
            description: away team id
          field:
            type: string
            example: JYSK park
            description: field of the games  
          referee:
            type: string
            example: Bobby Madden
            description: referee of the game  
          stage:
            type: string
            example: Championship Round
            description: stage of the game
...